Added fuzzyMatch.
Works both directly on the attributes of a record and attributes
of (nested) relation of the record.

Due to filtering on these relations, the change was a bit more
complex.

The adapter instance needs to be passed down the whole chain of
matching, since extra queries need to occur to walk the path of
relations.

Extra note
----------

It somewhat assumes the last leaf in the path refers to an attribute
of a record. If it ends in a relation, this won't provide the results you
want as a user.

This will work:
```
store.find('user', undefined, {
  fuzzyMatch: {
   'name': 'john D'
   'address:building:rooms:roomName': 'Room X'
   }
  }
)
```

This not:
```
store.find('user', undefined, {
  fuzzyMatch: {
   'name': 'john D'
   'address:building': 'Building Y'
   }
  }
)
```